Create Forms
Forms are essential components in HTML because the allow user input.
Among the most common input types we can find text, email, password, select, checkbox, and radio. 
Clicking on a label that's linked to an input will focus that input. 
Checkboxes and radio buttons can be grouped by using the same name attribute. 
The <textarea> element is used for large content, such as messages. 
In order to function properly, buttons should be placed inside the <form> element. Alternatively, you can use the form attribute on an external button to point to the correct form. 
Pressing the button that's inside a form or is linked to that form will execute a submit action.