Create DisplayingTabularDataWithTables
In the early days of the web, tables were being used to create layouts, Nowadays, this practice should be avoided because there are much better options: for example, Flexbox or CSS Grid Layout. 
Tables should be used solely for displaying tabular data. 
In a table, that <tr> element is used for table rows, the <th> for table headers and the <td> for table data cells. 
Optionally, you can use the <thead>, <tbody> and <tfoot> elements for wrapping various parts of the table contents.