Create CSSUnits
A lot of CSS properties have numeric values and most of those values use units.
On the web we can use various units to set various dimensions or create various layouts. 
Among the most commonly used units in CSS we find px, em, rem and %
px sets an absolute size which is not usually influenced by external factors. 
em is a relative unit that sets a size in relation to the font size of the parent element. 
The rem is also a relative unit, but it works in relation to the font size of the <html> element.
The % is also a relative unit and it represents a fraction of some other value. In CSS, that other value is usually a dimension of the parent element.