Create StylingTextInCSS
CSS offers a lot of control over how text looks like.
Font-family, font-size, and font-weight are properties that change the font you're using, the size of the text and the weight respectively. You can also use font-style to switch to regular, italic, or oblique. 
The text-transform CSS property specifies how to capitalize an element's text. you can choose between uppercase, lowercase and a few others. 
The letter-spacing property can be used to set the horizontal spacing between text characters. 
The letter-height property is used to set the height of a line box. In other words, it changes the distance between lines of text. 
you can align txt in different ways by using the text-align property. For values you have left (which is the default), center and right. You can also use start and end as substitutes for left and right. 
To change an element's appearance on mouse-over and mouse-press you can use the :hover and :active pseudo-classes. The :active pseudo-class is only applicable on a handful of elements (like buttons, links and certain inputs).