Create SettingDimensionsInCSS
You can change an elements width or height by using the width and, respectively, height properties. 
These accept the following value types: a length (expressed in px, em, rem and so on), a percentage or a keyword (for example: auto).
You can use the min-width and max-width to limit how short or long an element can be. The same is applicable for an element's height, by using the min-height and max-height properties.  
You cannot change the width or height of an inline element.