Create CSSBoxModel
Everything in CSS has a box around it.
That box is made up of other boxes, as defined by the CSS box model.
A typical element has the following structure, moving from the inside outwards: content box, padding box, border box and margin box. 
The box model doesn't fully apply to inline elements.
There are two types of box models: standard and alternative. 
If we set a width or height using the standard model, that width will only be applied to the content box. Any padding and borders will add to additional size to the visible element. 
If we set a width or height using the alternative model, that width will be applied to the entire visible element. Any padding and borders will be included in that size. 
The margin property will determine the size of the margin box. You can use the shorthand property to affect all box sides at once, or the longhand for setting each side individually. 
CSS accepts negative values for margin. 
If you have two elements with margins that touch, those margins will collapse to a single one. The size of the final margin depends on the size of the individual margins and also their sign. 
The padding property will determine the size of the padding box. Just like margins, you can use either the shorthand or the longhand properties. 
CSS only accepts positive values for padding.