Create WorkingWithBordersInCSS
You can control the size and appearance of the border box with the border properties. 
the border-width property controls the thickness of the width of an element's border. 
The border-color property defines the border color. 
The border-style property sets the line style for an element's border. We can remove the order altogether with the value of none, or we can choose from the following values: solid, dotted, dashed, double and a few others.
The border-radius property rounds the corners of an element's outer border edge.