Create DisplayProperty
Inner/Outer Display Type:
In CSS, boxes have an outer display type which specifies whether the box is block or inline. But they also have an inner display type that dictates how the elements inside the box are laid out. 


Notes for self:
In CSS, boxes have an outer display type which specifies whether the box is block or inline. But they also have an inner display type that dictates how the elements inside the box are laid out. 
The display type can be controlled through the display property. 
For the outer type, we can use values like none, block, inline and inline-block. 
The inline-block value allows an element to sit next to other elements, while also being able to accept a width and height and also display the top and bottom margin and padding correctly. 
For the inner type, we can use values like flex and grid, among a few others.