Create StylingListsInCSS
In CSS, lists can be styled using the following properties: list-style-image, list-style-position and list-style-type. Alternatively, use the shorthand list-style. 
The bullet point or number in a list item can be removed by setting the list-style or list-style-type to none. 
You can use an image as a list item marker with the list-style-image property. However, this is not recommended since it produces inconsistent results.