Update Positioning Elements in CSS
The position CSS property defines how an element is positioned. 
The final location of a positioned element is determined by the top, right, bottom, and left properties. 
By default, an element has the position value of static where the element is positioned according to the normal flow. The top, right, bottom and left properties have no effect. 
An element with position: relative is positioned according to the normal flow of the document, but then can be offset relative to itself by using the top, right, bottom and left properties. the offset does not affect the position of any other elements. 
An element with position: absolute is removed from the normal document flow and is positioned relative to its closest positioned ancestor. 
An element with position: fixed is also removed from the normal document flow and is positioned relative to the viewport. A fixed element stays in the same position as the user scrolls. 
An element with position: sticky is positioned based on the user's scroll position and will toggle between relative and fixed. The element is positioned relative until a given offset is met in the viewport. From that point its position is set to fixed. 
A sticky element stays within its parent so once the user scrolls past the parent, the sticky element is reverted back to a relative position.